Move auth related components to auth/
