+++ src/app/@authModal/(.)login/page.tsx
@@ -1,5 +1,5 @@
 import CloseModal from "@/components/CloseModal";
-import Login from "@/components/Login";
+import Login from "@/components/auth/Login";
 import { type FC } from "react";
 
 const page: FC = () => {
